Upload function CPost an article + Post a comment + Reply Comment

+++ SuViet/src/main/java/com/example/SuViet/repository/ArticleRepository.java
@@ -3,12 +3,11 @@ package com.example.SuViet.repository;
 import com.example.SuViet.model.Article;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-// import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
-public interface ArticleRepository extends JpaRepository<Article, Long> {
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
     Page<Article> findByEnabledTrue(Pageable pageable);
 
     // @Query("SELECT a FROM Article a WHERE a.enabled = true AND lower(a.title)
